Wallpaper updates when changed

+++ app/src/main/java/com/mini/infotainment/activities/maps/MapsActivity.kt
@@ -101,7 +101,7 @@ class MapsActivity : SActivity(), OnMapReadyCallback, MapInteractions {
     }
 
     private fun doOnLocationChanged(){
-        speedTw.text = gpsManager.currentSpeed.value.toString()
+        speedTw.text = gpsManager.currentSpeed.value.toInt().toString()
 
         if(googleMap == null) return
